Password change bug fixed
Hibernate manages entities related to a persistence context.
This means that all entity operations are intercepted by spring. This caused an
issue with changing password in UserService as YUHelper uses a session-based bean to
grab the current user. The means of grabbing this user are not through the persistence context
but rather the session context. Therefore, you need to grab the entity from the persistence context before
applying changes to it that you would like to persist. For some reason, though, changing the bean to a persistence-based user object will not work. Currently looking for a work around.

Some entity relations were also modified (nothing major, just adding cascade types)